Refactor Blender and Blender2 controllers to actions.
Also fixes a bug in handling of default filters in advanced search action.
