+++ module/VuFind/config/module.config.php
@@ -170,8 +170,6 @@ $config = [
             'VuFind\Controller\AlmaController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\AlphabrowseController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\DeveloperSettingsController' => 'VuFind\Controller\AbstractBaseFactory',
-            'VuFind\Controller\BlenderController' => 'VuFind\Controller\AbstractBaseFactory',
-            'VuFind\Controller\Blender2Controller' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\BrowseController' => 'VuFind\Controller\AbstractBaseWithConfigFactory',
             'VuFind\Controller\BrowZineController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\CartController' => 'VuFind\Controller\CartControllerFactory',
@@ -241,10 +239,6 @@ $config = [
             'alphabrowse' => 'VuFind\Controller\AlphabrowseController',
             'DeveloperSettings' => 'VuFind\Controller\DeveloperSettingsController',
             'developersettings' => 'VuFind\Controller\DeveloperSettingsController',
-            'Blender' => 'VuFind\Controller\BlenderController',
-            'blender' => 'VuFind\Controller\BlenderController',
-            'Blender2' => 'VuFind\Controller\Blender2Controller',
-            'blender2' => 'VuFind\Controller\Blender2Controller',
             'Browse' => 'VuFind\Controller\BrowseController',
             'browse' => 'VuFind\Controller\BrowseController',
             'BrowZine' => 'VuFind\Controller\BrowZineController',
